refactor: restructure data processors with inheritance-based architecture
- Replace prepare_json.py with modular BaseDataProcessor and LocalDataProcessor classes
- Extract common data processing logic into BaseDataProcessor base class
- Refactor UniProtQueryProcessor to inherit from BaseDataProcessor
- Move local data CLI functionality to dedicated cli/local_data.py module
- Update entry points and imports to reflect new module structure
- Improve code organization and reduce duplication across processors

Breaking change: rename protspace-json CLI command to protspace-local
