+++ pyproject.toml
@@ -63,7 +63,7 @@ dev-dependencies = [
 
 [project.scripts]
 protspace = "protspace.main:run"
-protspace-json = "protspace.data.prepare_json:main"
+protspace-local = "protspace.cli.local_data:main"
 protspace-query = "protspace.cli.uniprot_query:main"
 protspace-feature-colors = "protspace.utils.add_feature_style:main"
