+++ src/protspace/data/uniprot_query_processor.py
@@ -12,21 +12,19 @@ import numpy as np
 import pandas as pd
 from pymmseqs.commands import easy_search
 
-from protspace.utils import REDUCERS, DimensionReductionConfig
-from protspace.utils.reducers import MDS_NAME
+from protspace.utils import REDUCERS
+from protspace.data.base_data_processor import BaseDataProcessor
 from protspace.data.generate_csv import ProteinFeatureExtractor
 
 logging.basicConfig(format="%(levelname)s: %(message)s")
 logger = logging.getLogger(__name__)
 
-class UniProtQueryProcessor:
+class UniProtQueryProcessor(BaseDataProcessor):
     """Processor for UniProt query-based protein data analysis."""
-
-    REDUCERS = REDUCERS
 
     def __init__(self, config: Dict[str, Any]):
         # Remove command-line specific arguments that aren't used for dimension reduction
-        self.config = config.copy()
+        clean_config = config.copy()
         for arg in [
             "query",
             "sp",
@@ -41,9 +39,10 @@ class UniProtQueryProcessor:
             "with_fasta",
             "with_csv",
         ]:
-            self.config.pop(arg, None)
-        self.identifier_col = "identifier"
-        self.custom_names = config.get("custom_names", {})
+            clean_config.pop(arg, None)
+        
+        # Initialize base class with cleaned config and reducers
+        super().__init__(clean_config, REDUCERS)
 
     def process_query(
         self,
@@ -316,95 +315,4 @@ class UniProtQueryProcessor:
 
         return full_metadata
 
-    def process_reduction(self, data: np.ndarray, method: str, dims: int) -> Dict[str, Any]:
-        """Process a single reduction method."""
-        config = DimensionReductionConfig(n_components=dims, **self.config)
 
-        # Special handling for MDS when using similarity matrix
-        if method == MDS_NAME and config.precomputed is True:
-            # Convert similarity to dissimilarity matrix if needed
-            if np.allclose(np.diag(data), 1):
-                # Convert similarity to distance: d = sqrt(max(s) - s)
-                max_sim = np.max(data)
-                data = np.sqrt(max_sim - data)
-
-        reducer_cls = self.REDUCERS.get(method)
-        if not reducer_cls:
-            raise ValueError(f"Unknown reduction method: {method}")
-
-        reducer = reducer_cls(config)
-        reduced_data = reducer.fit_transform(data)
-
-        method_spec = f"{method}{dims}"
-        projection_name = self.custom_names.get(method_spec, f"{method.upper()}_{dims}")
-
-        return {
-            "name": projection_name,
-            "dimensions": dims,
-            "info": reducer.get_params(),
-            "data": reduced_data,
-        }
-
-    def create_output(
-        self,
-        metadata: pd.DataFrame,
-        reductions: List[Dict[str, Any]],
-        headers: List[str],
-    ) -> Dict[str, Any]:
-        """Create the final output dictionary."""
-        output = {"protein_data": {}, "projections": []}
-
-        # Process features
-        for _, row in metadata.iterrows():
-            protein_id = row[self.identifier_col]
-            features = (
-                row.drop(self.identifier_col)
-                .infer_objects(copy=False)
-                .fillna("")
-                .to_dict()
-            )
-            output["protein_data"][protein_id] = {"features": features}
-
-        # Process projections
-        for reduction in reductions:
-            projection = {
-                "name": reduction["name"],
-                "dimensions": reduction["dimensions"],
-                "info": reduction["info"],
-                "data": [],
-            }
-
-            for i, header in enumerate(headers):
-                coordinates = {
-                    "x": float(reduction["data"][i][0]),
-                    "y": float(reduction["data"][i][1]),
-                }
-                if reduction["dimensions"] == 3:
-                    coordinates["z"] = float(reduction["data"][i][2])
-
-                projection["data"].append(
-                    {"identifier": header, "coordinates": coordinates}
-                )
-
-            output["projections"].append(projection)
-
-        return output
-
-
-def save_output(data: Dict[str, Any], output_path: Path):
-    """Save output data to JSON file."""
-    if output_path.exists():
-        with output_path.open("r") as f:
-            existing = json.load(f)
-            existing["protein_data"].update(data["protein_data"])
-
-            # Update or add projections
-            existing_projs = {p["name"]: p for p in existing["projections"]}
-            for new_proj in data["projections"]:
-                existing_projs[new_proj["name"]] = new_proj
-            existing["projections"] = list(existing_projs.values())
-
-        data = existing
-
-    with output_path.open("w") as f:
-        json.dump(data, f, indent=2)
